+++ Src/Melville.Pdf.LowLevel/Parsing/ObjectParsers/IIndirectObjectResolver.cs
@@ -24,7 +24,7 @@ public interface IIndirectObjectRegistry
     /// <param name="number">Object number of the deleted block.</param>
     /// <param name="next">Object number of the next deleted block in the deleted block chain.</param>
     /// <param name="generation">Generation number of the </param>
-    void RegisterDeletedBlock(int number, ulong next, ulong generation);
+    void RegisterDeletedBlock(int number, int next, int generation);
     
     /// <summary>
     /// Registers an object of a type unknown to this version of the PDF parser.  This is either a
@@ -35,7 +35,7 @@ public interface IIndirectObjectRegistry
     /// <param name="number">The object number</param>
     /// <param name="next">A datum associated with this object</param>
     /// <param name="generation">The generation of the object</param>
-    void RegisterNullObject(int number, ulong next, ulong generation);
+    void RegisterNullObject(int number, int next, int generation);
     
     /// <summary>
     /// Register a PdfObject at a given position in the PDF file.
@@ -43,7 +43,7 @@ public interface IIndirectObjectRegistry
     /// <param name="number">The object number</param>
     /// <param name="generation">The generation number</param>
     /// <param name="offset">The offset from the %%PDF header to the beginning of this object.</param>
-    void RegisterIndirectBlock(int number, ulong generation, ulong offset);
+    void RegisterIndirectBlock(int number, int generation, long offset);
 
     /// <summary>
     /// Register a PDF object at a given position in a given object stream.
@@ -51,5 +51,5 @@ public interface IIndirectObjectRegistry
     /// <param name="number">The object in the stream</param>
     /// <param name="referredStreamOrdinal">Object number of the referred to stream.</param>
     /// <param name="positionInStream">The </param>
-    void RegisterObjectStreamBlock(int number, ulong referredStreamOrdinal, ulong positionInStream);
+    void RegisterObjectStreamBlock(int number, int referredStreamOrdinal, int positionInStream);
 }
